fix issue #53: switch from EditorActionBarContributor to ExtensionContributionFactory
