+++ Baselet/META-INF/MANIFEST.MF
@@ -7,9 +7,12 @@ Bundle-SymbolicName: com.umlet.plugin;singleton:=true
 Bundle-Version: 14.1
 Bundle-Activator: com.baselet.plugin.MainPlugin
 Bundle-Vendor: http://www.umlet.com
-Require-Bundle: org.eclipse.ui,org.eclipse.ui.ide,org.eclipse.core.res
- ources,org.eclipse.core.runtime,org.eclipse.jdt.core;resolution:=opti
- onal
+Require-Bundle: org.eclipse.ui,
+ org.eclipse.ui.ide,
+ org.eclipse.core.resources,
+ org.eclipse.core.runtime,
+ org.eclipse.jdt.core;resolution:=optional,
+ org.eclipse.core.expressions
 Bundle-ClassPath: ., lib/autocomplete.jar, lib/batik-awt-util.jar, lib
  /batik-dom.jar, lib/batik-ext.jar, lib/batik-svggen.jar, lib/batik-ut
  il.jar, lib/batik-xml.jar, lib/bcel-5.2.jar, lib/commons-io-2.4.jar,
