+++ Baselet/plugin.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<plugin>
+<plugin >
    <extension point="org.eclipse.ui.newWizards">
       <wizard
             name="Umlet diagram"
@@ -19,5 +19,10 @@
             extensions="uxf"
             default="false"
       />
+   </extension>
+   <extension point="org.eclipse.ui.menus">
+     <menuContribution
+            locationURI="menu:org.eclipse.ui.main.menu?after=additions"
+            class="com.baselet.plugin.gui.MenuContributor" />
    </extension>
 </plugin>
