+++ Baselet/src/com/baselet/plugin/gui/EclipseGUI.java
@@ -41,6 +41,8 @@ public class EclipseGUI extends BaseGUI {
 	private final HashMap<DiagramHandler, Editor> diagrams;
 	private Contributor contributor;
 
+	private MenuContributor menuContributor;
+
 	public EclipseGUI(CanCloseProgram main) {
 		super(main);
 		diagrams = new HashMap<DiagramHandler, Editor>();
@@ -65,7 +67,9 @@ public class EclipseGUI extends BaseGUI {
 			return; // Possible if method is called at loading a palette
 		}
 		boolean enable = handler != null && !currentDiagram.getGridElements().isEmpty();
-		contributor.setExportAsEnabled(enable);
+		if (menuContributor != null) {
+			menuContributor.setExportAsEnabled(enable);
+		}
 	}
 
 	@Override
@@ -300,8 +304,8 @@ public class EclipseGUI extends BaseGUI {
 
 	@Override
 	public void setValueOfZoomDisplay(int i) {
-		if (contributor != null) {
-			contributor.updateZoomMenuRadioButton(i);
+		if (menuContributor != null) {
+			menuContributor.updateZoomMenuRadioButton(i);
 		}
 	}
 
@@ -338,5 +342,9 @@ public class EclipseGUI extends BaseGUI {
 	@Override
 	public boolean saveWindowSizeInConfig() {
 		return false;
+	}
+
+	public void setMenuContributor(MenuContributor menuContributor) {
+		this.menuContributor = menuContributor;
 	}
 }
